Created Refactored Code for Main rover code
