+++ Individual_Component_Header_Files/Battery_Sensing_Header/main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "main.c" "blink.c"
+idf_component_register(SRCS "Battery_Sense.c" "main.c" "blink.c"
                     INCLUDE_DIRS ".")
